+++ packages/iov-encoding/src/index.ts
@@ -1,10 +1,21 @@
-export { fromAscii, toAscii } from "./ascii";
-export { fromBase64, toBase64 } from "./base64";
-export { Bech32 } from "./bech32";
-export { Decimal } from "./decimal";
-export { Encoding } from "./encoding";
-export { fromHex, toHex } from "./hex";
-export { Int53, Uint32, Uint53, Uint64 } from "./integers";
+// Re-exports for IOV Core 2.x compatibility
+export { Int53, Uint32, Uint53, Uint64, Decimal } from "@cosmjs/math";
+export {
+  fromHex,
+  toHex,
+  fromAscii,
+  toAscii,
+  fromBase64,
+  toBase64,
+  fromRfc3339,
+  toRfc3339,
+  fromUtf8,
+  toUtf8,
+  Bech32,
+} from "@cosmjs/encoding";
+export { isNonNullObject, isUint8Array } from "@cosmjs/utils";
+
+// Symbols not (yet) existent in CosmJS
 export {
   JsonCompatibleValue,
   isJsonCompatibleValue,
@@ -13,7 +24,7 @@ export {
   JsonCompatibleArray,
   isJsonCompatibleArray,
 } from "./json";
-export { fromRfc3339, toRfc3339 } from "./rfc3339";
 export { TransactionEncoder } from "./transactionencoder";
-export { isNonNullObject, isUint8Array } from "./typechecks";
-export { fromUtf8, toUtf8 } from "./utf8";
+
+// Deprecated symbols to be removed in the next mayor release
+export { Encoding } from "./encoding";